담당자 초대 관련 로직 수정 #125
